extract selected-translation-keys-editor
